feat: Take the public definition out of sc-dial folder and place it in the common folder
