+++ packages/synchro-charts/src/components/sc-dial/sc-dial-base/sc-dial-loading.tsx
@@ -1,5 +1,5 @@
 import { Component, h, Prop } from '@stencil/core';
-import { ColorConfigurations } from '../utils/util';
+import { ColorConfigurations } from '../../common/constants';
 
 const SPACE = 4;
 @Component({
