+++ packages/synchro-charts/src/components/sc-dial/utils/type.ts
@@ -1,3 +1,5 @@
+import { ErrorMessages, TooltipMessage } from '../../common/types';
+
 export interface DialSizeConfig {
   fontSize: number;
   dialThickness: number;
@@ -6,44 +8,8 @@ export interface DialSizeConfig {
   unitSize: number;
 }
 
-export type RecursivePartial<T> = {
-  [P in keyof T]?: T[P] extends (infer U)[]
-    ? RecursivePartial<U>[]
-    : T[P] extends object
-    ? RecursivePartial<T[P]>
-    : T[P];
-};
-
-export type DialErrorMessages = {
-  errorTimeLabel: string;
-  invalidValueError: string;
-  missingYminAndYmaxError: string;
-  missingYminError: string;
-  missingYmaxError: string;
-  invalidYminAndYmaxError: string;
-  dataNotLimitsError: string;
-};
-
-export type TooltipMessage = {
-  tooltipValueTitles: string;
-  tooltipValueTimeDescribed: string;
-  tooltipStatusTitles: string;
-  tooltipStatusDescribed: string;
-};
-
 export type DialMessages = {
-  error: DialErrorMessages;
+  error: ErrorMessages;
   loading: string;
   tooltip: TooltipMessage;
-};
-
-/**
- * These are radian.
- * For example:
- *    const angle = { startAngle: 0, endAngle: Math.PI } means `The range of angles is [0, 180°]`
- *    The display is a semi-circular arc.
- */
-export type AngleDefault = {
-  startAngle: number;
-  endAngle: number;
 };
